+++ admin/delete_comment_photo.php
@@ -10,7 +10,10 @@ if(empty($_GET['id']))
 $comment = Comment::find_by_id($_GET['id']);
 if($comment)
 {
-    $comment->delete();
+    if($comment->delete())
+    {
+        $comment->delete();
+    }
     redirect("comment_photo.php?id={$comment->photo_id}");
 }
 else
