+++ admin/delete_photo.php
@@ -8,9 +8,14 @@ if(empty($_GET['id']))
     redirect("photos.php");
 }
 $photo = Photo::find_by_id($_GET['id']);
+$user = User::find_by_id($_SESSION['user_id']);
+
 if($photo)
 {
-    $photo->delete_photo();
+    if($user->id == $photo->user_id || strpos(User::check_user_role(),'admin'))
+    {
+        $photo->delete_photo();
+    }
     redirect("photos.php");
 }
 else
